// dunadev-frontend/src/api/dunadev/model/locationSummary.ts
/**
 * DunaDev API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Minimal location info embedded in event responses.
 */
export interface LocationSummary { 
    id: number;
    name: string;
    address?: string | null;
    city?: string | null;
    latitude?: number | null;
    longitude?: number | null;
    websiteUrl?: string | null;
    howToGetThere?: string | null;
}