// dunadev-frontend/src/api/dunadev/model/eventRequest.ts
/**
 * DunaDev API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { EventLinkRequest } from './eventLinkRequest';


export interface EventRequest { 
    title: string;
    description?: string | null;
    /**
     * Link to the external event page.
     */
    eventUrl?: string | null;
    startsAt: string;
    endsAt?: string | null;
    free: boolean;
    registrationRequired: boolean;
    registrationUrl?: string | null;
    /**
     * The date-time from which the event becomes publicly visible. Omit or set to null to publish immediately. 
     */
    visibleFrom?: string | null;
    /**
     * ID of the organiser\'s location where the event takes place.
     */
    locationId: number;
    links?: Array<EventLinkRequest>;
}